Remove unnecessary type casts.
git-svn-id: https://svn.apache.org/repos/asf/jakarta/commons/proper/lang/trunk@137950 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/commons/lang/mutable/MutableInt.java
@@ -21,7 +21,7 @@ import java.io.Serializable;
  * A mutable <code>int</code>.
  * 
  * @since 2.1
- * @version $Id: MutableInt.java,v 1.2 2004/07/07 23:50:28 scolebourne Exp $
+ * @version $Id: MutableInt.java,v 1.3 2004/10/01 17:12:29 ggregory Exp $
  */
 public class MutableInt extends Number
         implements Comparable, Mutable, Serializable {
@@ -96,15 +96,15 @@ public class MutableInt extends Number
     }
 
     public long longValue() {
-        return (long) value;
+        return value;
     }
 
     public float floatValue() {
-        return (float) value;
+        return value;
     }
 
     public double doubleValue() {
-        return (double) value;
+        return value;
     }
 
     //-----------------------------------------------------------------------
